+++ gazebo/physics/SurfaceParams.hh
@@ -49,6 +49,22 @@ namespace gazebo
       /// \return Friction coefficient in secondary direction.
       public: double GetMuSecondary();
 
+      /// \brief Get the torsional friction coefficient.
+      /// \return Torsional friction coefficient.
+      public: double GetMuTorsion();
+
+      /// \brief Get the torsional patch radius.
+      /// \return Torsional patch radius.
+      public: double GetPatchRadius();
+
+      /// \brief Get the torsional surface radius.
+      /// \return Torsional surface radius.
+      public: double GetSurfaceRadius();
+
+      /// \brief Get the torsional "use patch radius" flag.
+      /// \return Torsional "use patch radius" flag.
+      public: bool GetUsePatchRadius();
+
       /// \brief Set the friction coefficient in the primary direction.
       /// \param[in] _mu Friction coefficient.
       public: void SetMuPrimary(double _mu);
@@ -56,6 +72,22 @@ namespace gazebo
       /// \brief Set the friction coefficient in the secondary direction.
       /// \param[in] _mu Friction coefficient.
       public: void SetMuSecondary(double _mu);
+
+      /// \brief Set the torsional friction coefficient.
+      /// \param[in] _mu Torsional friction coefficient.
+      public: void SetMuTorsion(double _mu);
+
+      /// \brief Set the torsional patch radius.
+      /// \param[in] _radius Torsional patch radius.
+      public: void SetPatchRadius(double _radius);
+
+      /// \brief Set the torsional surface radius.
+      /// \param[in] _radius Torsional surface radius.
+      public: void SetSurfaceRadius(double _radius);
+
+      /// \brief Set whether to use the surface radius.
+      /// \param[in] _use True to use the surface radius.
+      public: void SetUsePatchRadius(bool _use);
 
       /// \brief Get the friction coefficient in a single direction.
       /// \param[in] _index Index of friction direction, 0 for primary,
@@ -84,7 +116,26 @@ namespace gazebo
       /// \brief Array of dry friction coefficients. mu[0] is in the
       /// primary direction as defined by the friction pyramid.
       /// mu[1] is in the second direction.
-      private: double mu[2];
+      /// mu[2] is in the torsional direction.
+      private: double mu[3];
+
+      /// \brief Radius of the contact patch used to calculate torsional
+      /// friction. The same patch is used by all contact points in this
+      /// surface.
+      /// The patch is equal to the square root of the surface radius times the
+      /// contact depth: a = srqt(R * d).
+      private: double patchRadius;
+
+      /// \brief Radius of the surface to be used to calculate torsional
+      /// friction. The same radius is used for all contacts points in this
+      /// surface.
+      private: double surfaceRadius;
+
+      /// \brief Flag for choosing the method for computing the contact
+      /// patch radius in torsional friction.
+      /// True to use the constant patch radius parameter. False to use the
+      /// surface radius together with contact depth.
+      private: bool usePatchRadius;
     };
 
     /// \class SurfaceParams SurfaceParams.hh physics/physics.hh