+++ gazebo/physics/ode/ODEPhysics.cc
@@ -1061,11 +1061,14 @@ void ODEPhysics::Collide(ODECollision *_collision1, ODECollision *_collision2,
   // Set the contact surface parameter flags.
   contact.surface.mode = dContactBounce |
                          dContactMu2 |
+                         dContactMu3 |
                          dContactSoftERP |
                          dContactSoftCFM |
                          dContactApprox1 |
+                         dContactApprox3 |
                          dContactSlip1 |
-                         dContactSlip2;
+                         dContactSlip2 |
+                         dContactSlip3;
 
   ODESurfaceParamsPtr surf1 = _collision1->GetODESurface();
   ODESurfaceParamsPtr surf2 = _collision2->GetODESurface();
@@ -1130,13 +1133,28 @@ void ODEPhysics::Collide(ODECollision *_collision1, ODECollision *_collision2,
                                 surf2->GetFrictionPyramid()->GetMuPrimary());
   contact.surface.mu2 = std::min(surf1->GetFrictionPyramid()->GetMuSecondary(),
                                  surf2->GetFrictionPyramid()->GetMuSecondary());
-
+  contact.surface.mu3 = std::min(surf1->GetFrictionPyramid()->GetMuTorsion(),
+                                 surf2->GetFrictionPyramid()->GetMuTorsion());
+  contact.surface.patch_radius =
+      std::max(surf1->GetFrictionPyramid()->GetPatchRadius(),
+               surf2->GetFrictionPyramid()->GetPatchRadius());
+  // the curvature is combined using 1/R = 1/R1 + 1/R2
+  // we can consider doing the same for the patch radius
+  contact.surface.surface_radius = 1/
+      (1/surf1->GetFrictionPyramid()->GetSurfaceRadius()+
+      1/surf2->GetFrictionPyramid()->GetSurfaceRadius());
+  // not sure how to combine these logic flags
+  contact.surface.use_patch_radius =
+      surf1->GetFrictionPyramid()->GetUsePatchRadius() &&
+      surf2->GetFrictionPyramid()->GetUsePatchRadius();
 
   // Set the slip values
   contact.surface.slip1 = std::min(surf1->slip1,
                                    surf2->slip1);
   contact.surface.slip2 = std::min(surf1->slip2,
                                    surf2->slip2);
+  contact.surface.slip3 = std::min(surf1->slipTorsion,
+                                   surf2->slipTorsion);
 
   // Set the bounce values
   contact.surface.bounce = std::min(surf1->bounce,